chore: fix mock for deposit page test
